requirements.txt
azure-eventhub==5.10.1
azure-identity==1.12.0
openai==1.37.1
scikit-learn==1.2.2
pandas==2.0.3
numpy==1.24.3
kafka-python==2.0.2
requests==2.31.0
